[grid] Moved utils files inside /utils folder.
