+++ projects/angular-grid-layout/src/lib/utils/react-grid-layout.utils.ts
