+++ projects/angular-grid-layout/src/lib/utils/tests/react-grid-layout-utils.spec.ts
