+++ projects/angular-grid-layout/src/public-api.ts
@@ -1,7 +1,7 @@
 /*
  * Public API Surface of grid
  */
-export { ktdGridCompact, ktdTrackById } from './lib/grid.utils';
+export { ktdGridCompact, ktdTrackById } from './lib/utils/grid.utils';
 export * from './lib/directives/drag-handle';
 export * from './lib/directives/resize-handle';
 export * from './lib/grid-item/grid-item.component';
